+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 3.0-alpha10
 - Bug fixes:
   - -- Fixed error when adding an objective with no trackers to a bar
+  - -- Fixed errors preventing custom conditions from being set and used (may be reworked in the future)
 
 ### 3.0-alpha9
 - Reimplemented quick add editbox